Untangle legacy .gcroot field
When the new gc root placement pass was introduced, the .gcroot field
lost its original meaning and became mostly superfluous. However, the
codegen code for unions still used it to get a referenced to the boxed
part of the split union. Clean that up by introducing an explicit field
for that purpose and removing the .gcroot field.